+++ cola/hotkeys.py
@@ -81,6 +81,7 @@ TERMINAL = hotkey(Qt.CTRL | Qt.SHIFT | Qt.Key_T)
 STAGE_UNTRACKED = hotkey(Qt.ALT | Qt.Key_U)
 REVERT = hotkey(Qt.CTRL | Qt.Key_U)
 REVERT_ALT = hotkey(Qt.ALT | Qt.SHIFT | Qt.Key_R)
+REVERT_UNSTAGED_EDITS = hotkey(Qt.ALT | Qt.CTRL | Qt.Key_U)
 EDIT_AND_REVERT = hotkey(Qt.CTRL | Qt.SHIFT | Qt.Key_U)
 WORD_RIGHT = hotkey(Qt.Key_W)